Force promotion to long double

+++ src/util/rmatrix.c
@@ -516,7 +516,8 @@ rmx_multiply(const RMATRIX *m1, const RMATRIX *m2)
 	        for (k = mres->ncomp; k--; ) {
 		    long double	d = 0;
 		    for (h = m1->ncols; h--; )
-			d += rmx_lval(m1,i,h,k) * rmx_lval(m2,h,j,k);
+			d += (long double)rmx_lval(m1,i,h,k) *
+				(long double)rmx_lval(m2,h,j,k);
 		    rmx_lval(mres,i,j,k) = (double)d;
 		}
 	return(mres);
